backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ssahim-backend"
version = "0.1.0"
description = "sSsahim Arabic Investment Intelligence — backend API"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "httpx>=0.27.0",
    "openai>=1.30.0",
    "numpy>=1.26.0",
    "pandas>=2.2.0",
    "scipy>=1.13.0",
    "statsmodels>=0.14.0",
    "pydantic-settings>=2.2.0",
    "cachetools>=5.3.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "pytest-cov>=5.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
